# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Examplo API Do BI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/bi/v1/",
    "paths": {
        "/atualizarlinkbi": {
            "put": {
                "description": "Endpoint para atualizar Bi",
                "tags": [
                    "BI"
                ],
                "summary": "Atualizar link do BI",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id do seto responsavel",
                        "name": "idsetor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id do bi a ser atualizado ",
                        "name": "idbi",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "novo link do bi",
                        "name": "link",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/atualizarusuario": {
            "put": {
                "description": "Endpoint para atualizar usuario",
                "tags": [
                    "Usuario"
                ],
                "summary": "Atualizar usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id do setor responsavel",
                        "name": "usuario",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "novo nome do setor",
                        "name": "iduser",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "novo nome do setor",
                        "name": "idsetor",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/atualizasetor": {
            "put": {
                "description": "Endpoint para atualizar Setor",
                "tags": [
                    "Setor"
                ],
                "summary": "Atualizar Setor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id do setor responsavel",
                        "name": "id_setor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "novo nome do setor",
                        "name": "nome_setor",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/cadastrarbi": {
            "post": {
                "description": "Endpoint para cadastrar link do BI",
                "tags": [
                    "BI"
                ],
                "summary": "Cadastrar link do Bi",
                "parameters": [
                    {
                        "type": "string",
                        "description": "link do bi",
                        "name": "link",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id do setor",
                        "name": "id_setor",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/cadastrosetor": {
            "post": {
                "description": "Endpoint para cadastrar um novo setor",
                "tags": [
                    "Setor"
                ],
                "summary": "Cadastrar um novo setor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do setor a ser cadastrado",
                        "name": "nome",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/cadastrousuario": {
            "post": {
                "description": "Endpoint para cadastrar um novo usuario",
                "tags": [
                    "Usuario"
                ],
                "summary": "Cadastrar um novo usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do usuario a ser cadastrado",
                        "name": "nome",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "login do usuario",
                        "name": "login",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id do setor",
                        "name": "id_setor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "permisao",
                        "name": "permisao",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "email do usuario",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "senha do usuario",
                        "name": "senha",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/deletarlink": {
            "delete": {
                "description": "Endpoint para Deletar Link do Bi",
                "tags": [
                    "BI"
                ],
                "summary": "Deletar link do BI",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id do seto responsavel",
                        "name": "idlink",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/deletaruser": {
            "delete": {
                "description": "Endpoint para Deletar Usuario",
                "tags": [
                    "Usuario"
                ],
                "summary": "Deletar Usuario",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id do seto responsavel",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/deletesetor": {
            "delete": {
                "description": "Endpoint para Deletar Setor",
                "tags": [
                    "Setor"
                ],
                "summary": "Deletar Setor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id do seto responsavel",
                        "name": "idsetor",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/links": {
            "get": {
                "description": "Endpoint para GET usuarios",
                "tags": [
                    "BI"
                ],
                "summary": "Selecionar o link do BI",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id do setor",
                        "name": "idsetor",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/setor": {
            "get": {
                "description": "Endpoint para GET Setor",
                "tags": [
                    "Setor"
                ],
                "summary": "Selecionar o setor",
                "responses": {}
            }
        },
        "/usuarios": {
            "get": {
                "description": "Endpoint para GET usuarios",
                "tags": [
                    "Usuario"
                ],
                "summary": "Selecionar o usuarios",
                "responses": {}
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}